update icons, small style tweaks
